Some changes based on the feedback.
