+++ src/System.Private.ServiceModel/tests/Common/Scenarios/ServiceInterfaces.cs
@@ -83,12 +83,6 @@ public interface IWcfService
 
     [OperationContract]
     void ReturnContentType(string contentType);
-
-    [OperationContract(Action = "http://tempuri.org/IWcfService/GetAllEmployees")]
-    List<Employee> GetAllEmployees();
-
-    [OperationContract(Action = "http://tempuri.org/IWcfService/AddEmployee")]
-    void AddEmployee(Employee employee);
 }
 
 [ServiceContract]
@@ -534,4 +528,14 @@ public interface IVerifyWebSockets
 {
     [OperationContract()]
     bool ValidateWebSocketsUsed();
+}
+
+[ServiceContract]
+public interface IDataContractResolverService
+{
+    [OperationContract(Action = "http://tempuri.org/IDataContractResolverService/GetAllEmployees")]
+    List<Employee> GetAllEmployees();
+
+    [OperationContract(Action = "http://tempuri.org/IDataContractResolverService/AddEmployee")]
+    void AddEmployee(Employee employee);
 }
