+++ src/System.Private.ServiceModel/tools/IISHostedWcfService/App_code/IDataResolverService.cs
@@ -1,0 +1,23 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+
+using System;
+using System.Collections.Generic;
+using System.IO;
+using System.ServiceModel;
+using System.ServiceModel.Channels;
+
+namespace WcfService
+{
+    [ServiceContract]
+    public interface IDataContractResolverService
+    {
+        [OperationContract(Action = "http://tempuri.org/IDataContractResolverService/GetAllEmployees")]
+        List<Employee> GetAllEmployees();
+
+        [OperationContract(Action = "http://tempuri.org/IDataContractResolverService/AddEmployee")]
+        void AddEmployee(Employee employee);
+    }
+}
