+++ src/System.Private.ServiceModel/tools/IISHostedWcfService/App_code/IWcfService.cs
@@ -155,11 +155,5 @@ namespace WcfService
             ReplyAction = "http://www.contoso.com/IXmlMessageContarctTestService/EchoMessageResquestWithMessageHeaderResponse")]
         [XmlSerializerFormat(SupportFaults = true)]
         XmlMessageContractTestResponse EchoMessageResquestWithMessageHeader(XmlMessageContractTestRequestWithMessageHeader request);
-
-        [OperationContract(Action = "http://tempuri.org/IWcfService/GetAllEmployees")]
-        List<Employee> GetAllEmployees();
-
-        [OperationContract(Action = "http://tempuri.org/IWcfService/AddEmployee")]
-        void AddEmployee(Employee employee);
-    }
+     }
 }
