+++ src/System.Private.ServiceModel/tools/IISHostedWcfService/App_code/testhosts/DataContractResolverTestServiceHost.cs
@@ -18,7 +18,7 @@ namespace WcfService
         }
     }
 
-    public class DataContractResolverTestServiceHost : TestServiceHostBase<IWcfService>
+    public class DataContractResolverTestServiceHost : TestServiceHostBase<IDataContractResolverService>
     {
         protected override string Address { get { return "DataContractResolver"; } }
 
@@ -30,7 +30,7 @@ namespace WcfService
         protected override void ApplyConfiguration()
         {
             base.ApplyConfiguration();
-            this.AddServiceEndpoint(typeof(IWcfService), new BasicHttpBinding(), "");
+            this.AddServiceEndpoint(typeof(IDataContractResolverService), new BasicHttpBinding(), "");
             foreach (ServiceEndpoint endpoint in this.Description.Endpoints)
             {
                 foreach (OperationDescription operation in endpoint.Contract.Operations)
@@ -48,47 +48,6 @@ namespace WcfService
         public DataContractResolverTestServiceHost(Type serviceType, params Uri[] baseAddresses)
             : base(serviceType, baseAddresses)
         {
-        }
-    }
-}
-
-public class ManagerDataContractResolver : DataContractResolver
-{
-    private string Namespace
-    {
-        get { return typeof(Manager).Namespace ?? "global"; }
-    }
-
-    private string Name
-    {
-        get { return typeof(Manager).Name; }
-    }
-
-
-    public override Type ResolveName(string typeName, string typeNamespace, Type declaredType, DataContractResolver knownTypeResolver)
-    {
-        if (typeName == this.Name && typeNamespace == this.Namespace)
-        {
-            return typeof(Manager);
-        }
-        else
-        {
-            return knownTypeResolver.ResolveName(typeName, typeNamespace, declaredType, null);
-        }
-    }
-
-    public override bool TryResolveType(Type type, Type declaredType, DataContractResolver knownTypeResolver, out XmlDictionaryString typeName, out XmlDictionaryString typeNamespace)
-    {
-        if (type == typeof(Manager))
-        {
-            XmlDictionary dic = new XmlDictionary();
-            typeName = dic.Add(this.Name);
-            typeNamespace = dic.Add(this.Namespace);
-            return true;
-        }
-        else
-        {
-            return knownTypeResolver.TryResolveType(type, declaredType, null, out typeName, out typeNamespace);
         }
     }
 }
