+++ src/System.Private.ServiceModel/tools/IISHostedWcfService/Web.config
@@ -92,7 +92,7 @@
         <add factory="WcfService.WebSocketHttpsDuplexBinaryBufferedTestServiceHostFactory" service="WcfService.WSDuplexService" relativeAddress="~\WebSocketHttpsDuplexBinaryBuffered.svc" />
         <add factory="WcfService.WebSocketHttpsDuplexTextBufferedTestServiceHostFactory" service="WcfService.WSDuplexService" relativeAddress="~\WebSocketHttpsDuplexTextBuffered.svc" />
         <add factory="WcfService.WebSocketHttpVerifyWebSocketsUsedTestServiceHostFactory" service="WcfService.VerifyWebSockets" relativeAddress="~\WebSocketHttpVerifyWebSocketsUsed.svc" />
-        <add factory="WcfService.DataContractResolverTestServiceHostFactory" service="WcfService.WcfService" relativeAddress="~\DataContractResolver.svc" />
+        <add factory="WcfService.DataContractResolverTestServiceHostFactory" service="WcfService.DataContractResolverService" relativeAddress="~\DataContractResolver.svc" />
       </serviceActivations>
     </serviceHostingEnvironment>
   </system.serviceModel>
